Go with Tasmota platform-espressif32 2026.03.30 that includes Tasmota Arduino Core 3.1.11 based on IDF 5.3.4.20260127, joltwallet/littlefs v1.20.4 and littlefs-project/littlefs v2.11.2 Print out important log information regarding the filesystem.
