+++ platformio.ini
@@ -22,7 +22,8 @@ default_envs = esp32doit-devkit-v1-factory
 ;default_envs = wemos_lolin_s2_mini-factory
 
 [env]
-platform = https://github.com/tasmota/platform-espressif32/releases/download/2025.12.30/platform-espressif32.zip
+; Platform 2026.02.30 Tasmota Arduino Core 3.1.11 based on IDF 5.3.4.20260127, joltwallet/littlefs v1.20.4 and littlefs-project/littlefs v2.11.2
+platform = https://github.com/tasmota/platform-espressif32/releases/download/2026.03.30/platform-espressif32.zip
 framework = arduino
 board_build.app_partition_name = factory
 board_upload.offset_address = 0x10000 ; factory partition address in partitions for 4,8,16 MB
